Load *all* the .vimrc files, from the buffer all the way up to the git project root.
Rather than just loading:
        $ROOT/.vimrc

This will now let you load a series of .vimrc files, so that you can set
project-specific settings at the top and then over-ride or set other
more restricted/limited settings in individual sub-directories.

We find all the .vimrc files to load from the bottom-up, but then load
them top-down so that we can load the general stuff first and then
over-ride that with more specific things as we go.